# protocol: benchmarks for protocol parsing