--- a/protocol/instruction.go
+++ b/protocol/instruction.go
@@ -20,6 +20,7 @@ const MaxInstructionLength = 8192
 var (
 	ErrInstructionMissDot   = errors.New("instruction without dot")
 	ErrInstructionMissComma = errors.New("instruction without comma")
+	ErrInstructionMissSemi  = errors.New("instruction withou semi")
 	ErrInstructionBadDigit  = errors.New("instruction with bad digit")
 	ErrInstructionBadRune   = errors.New("instruction with bad rune")
 )
@@ -32,6 +33,61 @@ type Instruction struct {
 // NewInstruction ...
 func NewInstruction(elements []string) *Instruction {
 	return &Instruction{elements}
+}
+
+// ParseInstruction parses an instruction: 1.a,2.bc,3.def,10.abcdefghij;
+func ParseInstruction(raw []byte) (ins *Instruction, err error) {
+	var (
+		cursor   int
+		elements []string
+	)
+
+	bytes := len(raw)
+	for cursor < bytes {
+
+		// 1. parse digit
+		lengthEnd := -1
+		for i := cursor; i < bytes; i++ {
+			if raw[i]^'.' == 0 {
+				lengthEnd = i
+				break
+			}
+		}
+		if lengthEnd == -1 { // cannot find '.'
+			return nil, ErrInstructionMissDot
+		}
+		length, err := strconv.Atoi(string(raw[cursor:lengthEnd]))
+		if err != nil {
+			return nil, ErrInstructionBadDigit
+		}
+
+		// 2. parse rune
+		cursor = lengthEnd + 1
+		element := new(strings.Builder)
+		for i := 1; i <= length; i++ {
+			r, n := utf8.DecodeRune(raw[cursor:])
+			if r == utf8.RuneError {
+				return nil, ErrInstructionBadRune
+			}
+			cursor += n
+			element.WriteRune(r)
+		}
+		elements = append(elements, element.String())
+
+		// 3. done
+		if cursor == bytes-1 {
+			break
+		}
+
+		// 4. parse next
+		if raw[cursor]^',' != 0 {
+			return nil, ErrInstructionMissComma
+		}
+
+		cursor++
+	}
+
+	return NewInstruction(elements), nil
 }
 
 func (i Instruction) String() string {
@@ -98,60 +154,7 @@ func (io *InstructionIO) Read() (*Instruction, error) {
 	if err != nil {
 		return nil, err
 	}
-
-	var (
-		cursor   int
-		elements []string
-	)
-
-	// an instruction:
-	// 1.a,2.bc,3.def,10.abcdefghij;
-	bytes := len(raw)
-	for cursor < bytes {
-
-		// 1. parse digit
-		lengthEnd := -1
-		for i := cursor; i < bytes; i++ {
-			if raw[i]^'.' == 0 {
-				lengthEnd = i
-				break
-			}
-		}
-		if lengthEnd == -1 { // cannot find '.'
-			return nil, ErrInstructionMissDot
-		}
-		length, err := strconv.Atoi(string(raw[cursor:lengthEnd]))
-		if err != nil {
-			return nil, ErrInstructionBadDigit
-		}
-
-		// 2. parse rune
-		cursor = lengthEnd + 1
-		element := new(strings.Builder)
-		for i := 1; i <= length; i++ {
-			r, n := utf8.DecodeRune(raw[cursor:])
-			if r == utf8.RuneError {
-				return nil, ErrInstructionBadRune
-			}
-			cursor += n
-			element.WriteRune(r)
-		}
-		elements = append(elements, element.String())
-
-		// 3. done
-		if cursor == bytes-1 {
-			break
-		}
-
-		// 4. parse next
-		if raw[cursor]^',' != 0 {
-			return nil, ErrInstructionMissComma
-		}
-
-		cursor++
-	}
-
-	return NewInstruction(elements), nil
+	return ParseInstruction(raw)
 }
 
 // WriteRaw writes raw buffer into io output